fix : 그룹 가입 화면 진입 후 되돌아기 시나리오를 반복 시 크러쉬나는 이슈 해결
