+++ presentation/src/main/java/com/yapp/bol/presentation/view/group/join/GroupJoinViewModel.kt
@@ -95,7 +95,7 @@ class GroupJoinViewModel @Inject constructor(
         nickName: String,
         successValidateNickname: (Boolean) -> Unit,
     ) {
-        matchUseCase.getValidateNickName(groupId, nickName).collectLatest {
+        matchUseCase.getValidateNickName(groupId.toInt(), nickName).collectLatest {
             checkedApiResult(
                 apiResult = it,
                 success = { isAvailable ->
