+++ presentation/src/main/java/com/yapp/bol/presentation/view/group/search/GroupSearchFragment.kt
@@ -44,7 +44,7 @@ class GroupSearchFragment : BaseFragment<FragmentGroupSearchBinding>(R.layout.fr
             showJoinGroupDialog = {
                 binding.root.findNavController().navigateFragment(
                     R.id.action_groupSearchFragment_to_groupJoinFragment,
-                    "groupItem" to it,
+                    "groupId" to it.id,
                 )
             },
             changeButtonColor = {
